Create .gitignore
Sarai Hannah Ajai have created the .gitignore file, this file is used to specify intentionally untracked files or directories that Git should ignore when tracking changes in a repository. It's a way to exclude certain files or directories from being added to version control. .gitignore files are particularly useful in projects that use Git for version control, as they help manage what files should not be committed to the repository dated on 08-30-23.