Create iVoteBallot.ejs
The Developer, Sarai Hannah Aja written HTML codes for the iVoteBallot home page is responsible for creating the structure and content of the web page that users see when they visit the iVoteBallot website. The code includes various elements such as headings, paragraphs, images, and links that are used to convey information about the iVoteBallot services and allow users to interact with the site through hyperlinks. The HTML codes, also includes metadata such as title tags, meta descriptions, and meta keywords that provide information about the page's content to search engines and other web services. Overall, Sarai Hannah Ajai written HTML codes for the iVoteBallot home page plays a critical role in creating a user-friendly and informative website that effectively communicates the benefits of the iVoteBallot services to potential users dated on 04-22-23 and 08-30-23.