Create footer.ejs
The Developer, Sarai Hannah Ajai have written her HTML codes for the "footer.ejs" file, and in these HTML codes are responsible for creating the footer section of the webpage. And, it contains three main elements: two horizontal lines with class names, "hr_Red_line" and "hr_Blue_line," a list of social media links with class name "socialMedias_01," and a small text block with the copyright information.

The social media links are represented as a list of items with hyperlinks, and each item has an image of the corresponding social media icon with class name "footer_Image_01." Also, the copyright information is contained in a separate list with an id name, "iVoteBallot_Footer_Insignia_01."

Overall, Ms. Ajai have written her HTML codes language for the "footer.ejs" file which provides a visually appealing and functional footer section with easy access to social media links and important legal information for the iVoteBallot website dated on 04-22-23 and 08-30-23.