Create header.ejs
The Developer, Sarai Hannah Ajai had written her HTML codes for the header.ejs file is responsible for defining the header section of the HTML page that contains the logo, slogan, and navigation bar. In the given HTML code, the header.ejs file defines the layout and styling of the header section using flexbox containers. The .left_Column, .middle_Column, and .right_Column classes define the layout of the header's three columns that contain the logo, slogan, and navigation bar, respectively. The slogan_01 paragraph tag defines the slogan text that appears in the middle column of the header. The navigational_Bar_01 unordered list defines the navigation bar that appears in the right column of the header. The header.ejs file, also, includes image tags that define the icons used in the navigation bar. Finally, the hr_Red_line horizontal rule tag creates a red line below the header section to separate it from the rest of the page's content. Overall, Ms. Ajai written HTML codes for the header.ejs file plays an essential role in creating the visual and functional design of the iVoteBallot home page's header section dated on 04-22-23 and 08-30-23.