Create about.ejs
The Developer, Sarai Hannah Aja written HTML codes language for the about.ejs file to contains information about the iVoteBallot system. The file includes a title and a link to the iVoteBallot page, which is accessible through the URL http://localhost:8080/iVoteBallot. And, the page also contains a section titled "About," which is marked by a heading and separated from the rest of the content by a blue horizontal line. The main content of the page is structured using a flex container and includes a column titled "Our Mission." This column explains that in order to participate in iVoteBallot, individuals must have a valid
state identification card and provides information about how registration cards are distributed. Overall, Ms. Ajai had written the about.ejs file that provides a concise overview of the iVoteBallot system and its requirements dated on 04-23-23 and 08-30-23.