Create 404.ejs
Sarai Hannah Ajai have written her her 404.ejs file for the sole purpose to display an error message when a user requests a page that does not exist. The HTML code in the file is styled using CSS to ensure that the error message is presented in a clear and visually appealing manner. The CSS styling is applied to various elements in the file, including the title, headings, paragraphs, and links. And, Ms. Ajai overall aim is to provide users with a helpful message that explains why they are seeing the error and what they can do to resolve the issue dated on 04-23-23 and 08-30-23.