Create ivoteBallot.css
Sarai Hannah Ajai have written the given CSS code defines styles for the iVoteBallot_Title element and its links within the id iVoteBallot_Title element of the iVoteBallot webpage. The iVoteBallot_Title element is set to display as a flex container with center-aligned text and blue-colored font of variable size using the clamp() function. The links within the iVoteBallot_Title element have no underlines, transparent background, and blue-colored font. On hover, the links' background color is changed to red with white-colored font. The CSS also includes media queries that adjust the layout and padding of the iVoteBallot_Title element and its links for screens of different sizes dated on 04-23-23 and 08-30-23.