Create 404.css
Sarai Hannah Ajai have written her CSS code language within the 404.css file has multiple purposes for formatting the iVoteBallot webpage's appearance when an HTTP 404 error occurs. It defines a custom font, sets the styles for the body element, and styles various elements using the flexbox layout. The file sets up a flexible container with a red background color and aligns the content center to help display information. The styles also include setting font size and color, defining a horizontal line with a solid border, and setting up a responsive and adaptive layout for web page elements. Overall, the 404.css file's main purpose is to ensure consistency and readability of the error page by formatting its appearance dated on 04-23-23 and 08-30-23.