Create 535.ejs
Sarai Hannah Ajai have written her HTML codes language for the 535.ejs file which provides information about the SMTP Error 535: Authentication Failed. It explains that when an user use the Simple Mail Transfer Protocol (SMTP) to send emails, a server generates an error message. And, if it detects inappropriate authentication by an user. The 535.ejs file lists several possible reasons why this error may occur, including incorrect username or password, disabled account, connection encryption, or unsupported authentication methods. It also provides a link to return to the iVoteBallot homepage and contact information if the user continues to experience the error. Finally, the file includes common header and footer code using the EJS templating engine to maintain consistency across the website dated on 04-24-23 and 08-30-23.