Create contactUs_01.ejs
Sarai Hannah Aja had written her HTML and JavaScript coded language for the contactUs_01.ejs file that includes a form for user to fill in and submit his/her contact us form to the iVoteBallot website. And, the contactUs_01.ejs file includes fields for the user's first name, middle name, last name, email address, confirmation of email address, whether he/she have an iVoteBallot account, and his/her concern. The contactUs_01.ejs file also, includes a textarea for the user to write their message and press the submit button; in order to submit his/her contact us form.

This script tag defines two functions: countCharacters and validateContactUsForm:

The countCharacters function retrieves the length of the text entered into a
contact us form's textMessage element and updates the textareaRemainingCharacters element to display the remaining number of characters available for input or a message indicating that the maximum number of characters has been reached.

The validateContactUsForm function retrieves the values entered into several
contact us form elements than performs a series of validation checks on each value, and displays an error message, if a value is invalid. The function also sets the retrieved values in the browser's session storage dated on 04-29-23 and 08-30-23.