Create 401.ejs
Sarai Hannah Ajai have written her HTML coded language for the 401.ejs error page for which is designed to inform users that they have encountered a 401 Unauthorized Access Error, which is part of the 4XX error code family. This error message indicates that there is an issue with the client's request, specifically when the server is unable to recognize the client. This type of error typically occurs due to incorrect login credentials, attempting to access a restricted page without authentication, or using an expired token or session ID. It can also occur when accessing iVoteBallot website or application from a new device that has not yet been authenticated. The 401.ejs error page provides users with information about the error and guidance on how to resolve it dated on 04-24-23 and 08-30-23.