Create alabama_VerifyTemporaryPwd_01.ejs
The Developer, Sarai Hannah Ajai have written her HTML and JavaScript codes languages for
the purpose of the alabama_VerifyTemporaryPwd_01.ejs web page is designed for users in the
State of Alabama to verify their temporary password sent by email verifications. It
consists of a form where users can enter their email address and temporary password. The
alabama_VerifyTemporaryPwd_01.ejs web page includes validation functions for the email
address and temporary password fields to ensure they are entered correctly. There is also
a JavaScript function that allows users to toggle the visibility of the password field.
And, the alabama_VerifyTemporaryPwd_01.ejs web page includes links for password recovery
and account creation dated on 08-30-23.