Create alabama_ForgotPassword_01.ejs
The Developer, Sarai Hannah Ajai have written her HTML and JavaScript codes languages for the
"alabama_ForgotPassword_01.ejs" file web page that serves, as a front-end interface for users     who have forgotten their password onto the iVoteBallot's website or web application accounts.     The alabama_ForgotPassword_01.ejs web page consists of various elements and functionalities to     facilitate the password recovery process.

At the top of the alabama_ForgotPassword_01.ejs web page, there is a common header section     included using the ('../common/header') code, which provides consistent branding and navigation options for the iVoteballot's website or web application.

The main content of the ForgotPassword_Alabama_01.ejs web page is contained within a <div>     element with the ID "flex_Container." Inside this container, there are multiple nested <div> elements that structure the layout of the alabama_ForgotPassword_01.ejs web page. The "first_Column" <div> contains the form for the password recovery process.

Within the form, there is a pop-up menu element with the ID "popUpForgotPasswordMenu" that is used to display additional options for the password recovery process.

The form itself has the class "form_ForgotPassword_01" and is set to submit data to the    "/verifyTempPwdLogIn_Alabama_01" URL using the POST method. It includes input fields for the     user's email address and a submit button labeled "Continue."

Below the email input field, there are two sections with hyperlinks. The first section asks, if the user is already a member of iVoteBallot website or web application and provides a link to access their account. The second section is for new users and offers a link to sign up for an iVoteBallot account.

At the bottom of the page, a common footer section is included using the   include('../common/footer') code, providing additional information and links.

Lastly, there is a script section at the end of the page that contains JavaScript coded language. It defines a function named "validateForgotPasswordForm" that is responsible for validating the email input field before submitting the form. The function checks, if the email field is not empty, has a minimum length of 6 characters, and matches a regular expression pattern for a valid email address format. If any validation errors occur, an error message is displayed. The script also registers an event listener to handle form submission and calls the "validateForgotPasswordForm" function dated 08-30-23.