Create header.css
Sarai Hannah Ajai have created the header.ejs file which is a template file used for rendering the header section of a webpage. It includes HTML and CSS code to define the layout and styling of the header. The header typically contains elements such as a navigation menu, logo, and search bar. The purpose of the header is to provide users with easy access to important information and functionality on the website dated on 04-22-23 and 09-01-23.