Create about.css
Sarai Hannah Ajai have written her about.css file; in order to contains a set of CSS style rules that define the appearance of various elements on the about.ejs webpage. The first rule defines a custom font called "Megrim-Regular" and sets its source file in the "src" property. It also sets the font weight to "normal." The next two rules style horizontal lines with blue and red borders, respectively, and set margins to visually separate sections on the webpage. The fourth rule styles the "iVoteBallot_Title" with a flexbox layout, using the custom font, adjusting font size based on the viewport, and setting a blue color for the text. The next three rules style the links in the "iVoteBallot_Title" with no underline decoration, transparent background, and blue text color. She, also add a red background and white text color on hover for visual interaction. The next two rules adjust the layout of the "iVoteBallot_Title" for smaller screens, distributing items with equal space and displaying in a vertical column with no padding. The last rule styles the "iVoteBallot_About_Title" with centered text alignment, the custom font, a dynamic font size, and a red font color, vertically centering it with margin-top and margin-bottom, and centering it horizontally with justify-content dated on 04-23-23 and 09-01-23.