Create views_Controller.js
Sarai Hannah Ajai written JavaScript coded language defines several functions that serve as controllers for rendering views to the iVoteBallot web application. These controllers are designed to work with the Model-View-Controller (MVC) architecture requirements, which separates an iVoteBalot application's data (the model), its presentation (the view), and the logic that connects the two to the (the controller).

The views_Controllers.js file requires a views_Router module and a path module. It then defines several controller functions that each render a specific view when called. The views include a custom 401 error page, a custom 404 error page, a custom 500 error page, a custom 535 error page, an "about" page, a "contact us" page, "iVote" ballot page and etc.

Finally, the module.exports statement makes these controller functions available to other parts of the iVoteBallot web application that require them dated on 04-26-23 and 09-01-23