Update index.js
This line of code Sarai Hannah Ajai (the Developer) had written in JavaScript language is used to create a Stripe client object in a JavaScript within the Node.js application. Stripe is a popular payment processing platform, and this code initializes a client for interacting with Stripe's services.

Here's a breakdown of how this line of code is executed:

1. const stripe: This declares a constant variable named "stripe." It will be used to reference the Stripe client object once it's created.

2. require('stripe'): This is a Node.js function for importing the "stripe" module. The "stripe" API library/module provides a JavaScript API for interacting with Stripe's services.

3. (process.env.STRIPE_PRIVATE_KEY): This part of the code accesses an environment variable named "STRIPE_PRIVATE_KEY." Environment variables are typically used to store sensitive information like API keys, so they are not hard-coded within the iVoteBallot_Prototype_3 web application. In this case, "STRIPE_PRIVATE_KEY" is to contain the iVoteBallot_Prototype_3's Stripe private API key.

4.  The (process.env.STRIPE_PRIVATE_KEY) is passed as an argument to the "stripe" module import. This means that the "stripe" module is being initialized with the iVoteBallot_Prototype_3 private key, allowing it to make authenticated requests to Stripe's servers on Ms. Ajai behalf.

So, when this line of code is executed, it results in a "stripe" client object that is configured with  the iVoteBallot_Prototype_3 key, and Ms. Ajai can use this object to make API calls to Stripe's services, such as creating charges, managing customers, or performing other payment-related tasks dated on 09-15-23.