Update views_Controller.js
Sarai Hannah Ajai written JavaScript coded language defines several functions that serve as controllers for rendering views to the iVoteBallot web application. These controllers are designed to work with the Model-View-Controller (MVC) architecture requirements, which separates an iVoteBalot application's data (the model), its presentation (the view), and the logic that connects the two to the (the controller).

The views_Controllers.js file requires a views_Router module and a path module. It then defines the controller function that render a specific view when called. The view include a custom "view_Alabama_Create_Checkout_Session".

Finally, the module.export statement makes these controller functions available to other parts of the iVoteBallot web application that require them dated on 10-07-23.