Update views_Router.js
Sarai Hannah Ajai have written her JavaScript coded language; in order to creates an instance of the Express routers and defines various routes with corresponding controller functions for handling views in a Model-View-Controller (MVC) architecture.

The routes include URL or displaying the web pages for "view_Alabama_Create_Checkout_Session.

This route is associated with a specific function defined in the views_Controller
module that is responsible for rendering the appropriate view.

Finally, the module exports and the router instances are for use, in other parts
of the iVoteBallot web application dated on 10-07-23.