Update package-lock.json
Sarai Hannah Ajai had added the following cors, stripe, and yaml libraries/modules to the "package.json" and package-lock.json files. And, the `package.json` file is a fundamental component of Node.js projects managed by the Node Package Manager (npm), and it plays a pivotal role in the development and distribution of your code. When working with Visual Studio Code, understanding and utilizing the `package.json` file is essential. And, the package-lock.json from the VSC terminal using the "npm init", and the `package-lock.json` is a crucial file used in Node.js projects managed by the Node Package Manager (npm), especially when working with Visual Studio Code (VSC). It serves as a detailed record of the project's dependencies and their exact versions, ensuring consistent and reproducible builds across different environments dated on 10-27-23.