Update alabama_SignUp_01_Controller.js
Sarai Hannah Ajai have added the middleware to set req.isUnauthenticated for the users' first use of the '/alabama_SignUp_01' URL bar. And, Ms. Ajai have added the constant redirectDashboard is a middleware function that checks if the user is already logged in by verifying the existence of a userId property in the user's session. If the user is logged in, the function redirects them to the dashboard page; otherwise, it allows the request to proceed to the next middleware function. This middleware is commonly used to restrict access to certain routes for authenticated users dated on 11-02-23.